Fix hashCode contract violations
hashCode() contract:
If two objects are equal according to equals(), their hashCode() must
produce the same integer result.

Location is an entity. It's identity is fully determined by its ID. Two
Location objects are equal (representing the same entity) if they have
the same ID. Other properties do not affect Location's identity.
Therefore, other properties are not taken into account in equals() and
must not affect hashCode().

+++ optaweb-vehicle-routing-backend/src/main/java/org/optaweb/vehiclerouting/domain/Coordinates.java
@@ -73,7 +73,7 @@ public class Coordinates {
 
     @Override
     public int hashCode() {
-        return Objects.hash(latitude, longitude);
+        return Objects.hash(latitude.doubleValue(), longitude.doubleValue());
     }
 
     @Override
